Moved logging and utils packages into their own modules
