+++ maverick-logging/.gitignore
@@ -1,0 +1,2 @@
+/target/
+/.classpath
