+++ maverick-logging/src/main/java/com/sshtools/common/logger/ConsoleLoggingContext.java
