+++ maverick-logging/src/main/java/com/sshtools/common/logger/RootLoggerContext.java
