+++ maverick-utils/src/main/java/com/sshtools/common/util/DynamicBuffer.java
