+++ maverick-utils/src/main/java/com/sshtools/common/util/EOLProcessor.java
