+++ maverick-utils/src/main/java/com/sshtools/common/util/SimpleASNWriter.java
