+++ maverick-utils/src/main/java/com/sshtools/common/util/URLUTF8Encoder.java
